Move getAddress to BlazeClientBuilder
