+++ blaze-client/src/main/scala/org/http4s/client/blaze/Http1Support.scala
@@ -19,7 +19,6 @@ package client
 package blaze
 
 import cats.effect._
-import cats.syntax.all._
 import java.net.InetSocketAddress
 import java.nio.ByteBuffer
 import java.nio.channels.AsynchronousChannelGroup
@@ -52,7 +51,7 @@ final private class Http1Support[F[_]](
     userAgent: Option[`User-Agent`],
     channelOptions: ChannelOptions,
     connectTimeout: Duration,
-    customDnsResolver: Option[RequestKey => Either[Throwable, InetSocketAddress]]
+    getAddress: RequestKey => Either[Throwable, InetSocketAddress]
 )(implicit F: ConcurrentEffect[F]) {
   private val connectionManager = new ClientChannelFactory(
     bufferSize,
@@ -65,7 +64,7 @@ final private class Http1Support[F[_]](
 ////////////////////////////////////////////////////
 
   def makeClient(requestKey: RequestKey): F[BlazeConnection[F]] =
-    customDnsResolver.getOrElse(getAddress _)(requestKey) match {
+    getAddress(requestKey) match {
       case Right(a) => fromFuture(F.delay(buildPipeline(requestKey, a)))
       case Left(t) => F.raiseError(t)
     }
@@ -128,11 +127,4 @@ final private class Http1Support[F[_]](
     }
   }
 
-  private def getAddress(requestKey: RequestKey): Either[Throwable, InetSocketAddress] =
-    requestKey match {
-      case RequestKey(s, auth) =>
-        val port = auth.port.getOrElse(if (s == Uri.Scheme.https) 443 else 80)
-        val host = auth.host.value
-        Either.catchNonFatal(new InetSocketAddress(host, port))
-    }
 }
